test: add upload test to CI
