+++ hooks/lib/upload.bash
@@ -1,9 +1,6 @@
-#[[ "${BASH_SOURCE[0]}" == "${0}" ]] && echo "Not for direct execution" && exit 2 || true
+[[ "${BASH_SOURCE[0]}" == "${0}" ]] && echo "Not for direct execution" && exit 2 || true
 _lib_script_dir="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
 set -euo pipefail
-
-# TODO
-set -x
 
 # shellcheck source=./run.bash
 source "$_lib_script_dir/run.bash"
